Change string interfaceis' parameters
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/469
Tested-by: Jenkins Server

+++ src/include/string.h
@@ -29,12 +29,11 @@
 extern "C" 
 {
 #endif
-
-    void *memset(void* s, int64_t c, size_t n);
+    void *memset(void* s, int c, size_t n);
     void bzero(void *vdest, size_t len);
     void *memcpy(void *dest, const void *src, size_t num);
     void *memmove(void *vdest, const void *vsrc, size_t len);
-    int64_t memcmp(const void *p1, const void *p2, size_t len);
+    int memcmp(const void *p1, const void *p2, size_t len);
 
     char* strcpy(char* d, const char* s);
     int strcmp(const char* s1, const char* s2);
